pulls out fetchAndSetMeta, cleanup
