+++ operations/getStation.ts
